solve the carsh coused by overried hash/refine if statements with guard (#3)
1.solve the carsh coused by overried hash
2. replace the if ststement with guard
3. add tests for DIDstore

+++ DID.Swift.SDK/ElastosDIDSDK/DID/Util/HDKey.swift
@@ -35,8 +35,8 @@ public class HDKey: NSObject {
     }
     
     // DerivedKey 初步推断是bip44 对于BitcoinKit的HDKeychain
-    public func derive(_ index: Int32) throws -> DerivedKey {
-        return DerivedKey(seed, index)
+    public func derive(_ index: Int) throws -> DerivedKey {
+        return DerivedKey(seed, Int32(index))
     }
